+++ include/rnetlib/verbs/verbs_local_memory_region.h
@@ -52,9 +52,9 @@ class VerbsLocalMemoryRegion : public LocalMemoryRegion {
 
   size_t GetLength() const override { return mr_ ? mr_->length : 0; }
 
-  uint32_t GetLKey() const override { return mr_ ? mr_->lkey : 0; }
+  uint64_t GetLKey() const override { return mr_ ? mr_->lkey : 0; }
 
-  uint32_t GetRKey() const override { return mr_ ? mr_->rkey : 0; }
+  uint64_t GetRKey() const override { return mr_ ? mr_->rkey : 0; }
 
  private:
   struct ibv_mr *mr_;
